+++ nodejs/awsx/ecs/fargateService.ts
@@ -231,7 +231,8 @@ export class FargateService extends ecs.Service {
     }
 }
 
-function getSubnets(
+/** @internal */
+export function getSubnets(
         cluster: ecs.Cluster,
         subnets: pulumi.Input<pulumi.Input<string>[]> | undefined,
         assignPublicIp: pulumi.Output<boolean>) {
@@ -343,16 +344,13 @@ type OverwriteFargateServiceArgs = utils.Overwrite<ecs.ServiceArgs, {
     taskDefinitionArgs?: FargateTaskDefinitionArgs;
     launchType?: never;
     networkConfiguration?: never;
+    capacityProviderStrategies?: never;
+    orderedPlacementStrategies?: never;
     securityGroups?: x.ec2.SecurityGroupOrId[];
 }>;
 
 export interface FargateServiceArgs {
     // Properties from ecs.ServiceArgs
-
-    /**
-     * The capacity provider strategy to use for the service.
-     */
-    capacityProviderStrategies?: aws.ecs.ServiceArgs["capacityProviderStrategies"];
 
     /**
      * onfiguration block containing deployment controller configuration.
@@ -436,13 +434,6 @@ export interface FargateServiceArgs {
      * is false, then these will be the private subnets of the cluster's vpc.
      */
     subnets?: pulumi.Input<pulumi.Input<string>[]>;
-
-    /**
-     * Service level strategy rules that are taken into consideration during task placement. List
-     * from top to bottom in order of precedence. The maximum number of `ordered_placement_strategy`
-     * blocks is `5`. Defined below.
-     */
-    orderedPlacementStrategies?: aws.ecs.ServiceArgs["orderedPlacementStrategies"];
 
     /**
      * rules that are taken into consideration during task placement. Maximum number of